# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mobile Money Fraud Detection API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TransactionType =
  (typeof TransactionType)[keyof typeof TransactionType];

export const TransactionType = {
  send: "send",
  receive: "receive",
  withdraw: "withdraw",
  agent_deposit: "agent_deposit",
  agent_withdrawal: "agent_withdrawal",
  airtime: "airtime",
  bill_payment: "bill_payment",
} as const;

export type TransactionFraudLabel =
  (typeof TransactionFraudLabel)[keyof typeof TransactionFraudLabel];

export const TransactionFraudLabel = {
  clean: "clean",
  suspicious: "suspicious",
  fraudulent: "fraudulent",
} as const;

export type FraudAnalysisFraudLabel =
  (typeof FraudAnalysisFraudLabel)[keyof typeof FraudAnalysisFraudLabel];

export const FraudAnalysisFraudLabel = {
  clean: "clean",
  suspicious: "suspicious",
  fraudulent: "fraudulent",
} as const;

export type FeatureImportanceDirection =
  (typeof FeatureImportanceDirection)[keyof typeof FeatureImportanceDirection];

export const FeatureImportanceDirection = {
  increases_risk: "increases_risk",
  decreases_risk: "decreases_risk",
  neutral: "neutral",
} as const;

export interface FeatureImportance {
  feature: string;
  value: number;
  importance: number;
  direction: FeatureImportanceDirection;
  description: string;
}

export interface FraudAnalysis {
  /** Risk score 0-100 */
  riskScore: number;
  isFraud: boolean;
  fraudLabel: FraudAnalysisFraudLabel;
  triggeredRules: string[];
  featureImportance: FeatureImportance[];
  explanation: string;
  modelUsed: string;
  confidence: number;
}

export interface Transaction {
  id: string;
  type: TransactionType;
  amount: number;
  currency: string;
  senderPhone: string;
  recipientPhone?: string;
  senderLocation: string;
  deviceId: string;
  timestamp: string;
  /** Account age in days */
  accountAge: number;
  previousTxCount: number;
  riskScore: number;
  isFraud: boolean;
  fraudLabel: TransactionFraudLabel;
  fraudAnalysis?: FraudAnalysis;
  createdAt: string;
}

export type CreateTransactionBodyType =
  (typeof CreateTransactionBodyType)[keyof typeof CreateTransactionBodyType];

export const CreateTransactionBodyType = {
  send: "send",
  receive: "receive",
  withdraw: "withdraw",
  agent_deposit: "agent_deposit",
  agent_withdrawal: "agent_withdrawal",
  airtime: "airtime",
  bill_payment: "bill_payment",
} as const;

export interface CreateTransactionBody {
  type: CreateTransactionBodyType;
  amount: number;
  currency?: string;
  senderPhone: string;
  recipientPhone?: string;
  senderLocation: string;
  deviceId: string;
  accountAge?: number;
  previousTxCount?: number;
  hourOfDay?: number;
  isNewDevice?: boolean;
  distanceFromHome?: number;
}

export type FraudAlertSeverity =
  (typeof FraudAlertSeverity)[keyof typeof FraudAlertSeverity];

export const FraudAlertSeverity = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface FraudAlert {
  id: string;
  transactionId: string;
  severity: FraudAlertSeverity;
  riskScore: number;
  amount: number;
  senderPhone: string;
  type: string;
  timestamp: string;
  explanation: string;
  triggeredRules: string[];
}

export type SimulatorConfigAmountRange = {
  min: number;
  max: number;
};

export interface SimulatorConfig {
  count: number;
  /** Fraction 0-1 of transactions to be fraudulent */
  fraudRate?: number;
  location?: string;
  amountRange?: SimulatorConfigAmountRange;
  transactionTypes?: string[];
  includeAgentTransactions?: boolean;
}

export interface SimulationSummary {
  total: number;
  flagged: number;
  fraudulent: number;
  avgRiskScore: number;
  totalAmount: number;
}

export type DashboardStatsByTypeItem = {
  type: string;
  count: number;
  fraudCount: number;
};

export type DashboardStatsByLocationItem = {
  location: string;
  count: number;
  fraudCount: number;
};

export interface DashboardStats {
  totalTransactions: number;
  flaggedTransactions: number;
  fraudulentTransactions: number;
  totalAmount: number;
  avgRiskScore: number;
  fraudRate: number;
  recentAlerts: number;
  blockedAmount: number;
  byType: DashboardStatsByTypeItem[];
  byLocation: DashboardStatsByLocationItem[];
}

export interface ModelMetrics {
  precision: number;
  recall: number;
  f1Score: number;
  accuracy: number;
  specificity: number;
  falsePositiveRate: number;
  falseNegativeRate: number;
  truePositives: number;
  trueNegatives: number;
  falsePositives: number;
  falseNegatives: number;
  auc: number;
}

export interface RocPoint {
  fpr: number;
  tpr: number;
  threshold: number;
}

export interface VolumeDataPoint {
  time: string;
  total: number;
  fraudulent: number;
  suspicious: number;
  clean: number;
}

export type ModelConfigActiveModels = {
  ruleBased: boolean;
  anomalyDetection: boolean;
  logisticRegression: boolean;
};

export type ModelConfigThresholds = {
  /** Score above which a transaction is suspicious */
  suspicious: number;
  /** Score above which a transaction is fraudulent */
  fraudulent: number;
  /** Score above which a transaction is blocked */
  block: number;
};

export type ModelConfigWeights = {
  ruleBased: number;
  anomalyDetection: number;
  logisticRegression: number;
};

export type ModelConfigFeatureFlags = {
  velocityCheck: boolean;
  geolocationCheck: boolean;
  deviceFingerprint: boolean;
  accountAgeCheck: boolean;
  amountAnomaly: boolean;
  timePattern: boolean;
};

export interface ModelConfig {
  activeModels: ModelConfigActiveModels;
  thresholds: ModelConfigThresholds;
  weights: ModelConfigWeights;
  featureFlags: ModelConfigFeatureFlags;
}

export type ListTransactionsParams = {
  limit?: number;
  offset?: number;
  flaggedOnly?: boolean;
};

export type ListTransactions200 = {
  transactions: Transaction[];
  total: number;
};

export type GenerateTransactions200 = {
  transactions: Transaction[];
  summary: SimulationSummary;
};

export type ResetSimulator200 = {
  message: string;
};

export type GetFraudAlertsParams = {
  limit?: number;
};

export type GetFraudAlerts200 = {
  alerts: FraudAlert[];
};

export type GetRocCurve200 = {
  points: RocPoint[];
  auc: number;
};

export type GetTransactionVolume200 = {
  data: VolumeDataPoint[];
};