# lib/api-zod/src/generated/types/featureImportance.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mobile Money Fraud Detection API
 * OpenAPI spec version: 0.1.0
 */
import type { FeatureImportanceDirection } from "./featureImportanceDirection";

export interface FeatureImportance {
  feature: string;
  value: number;
  importance: number;
  direction: FeatureImportanceDirection;
  description: string;
}