# lib/api-zod/src/generated/types/fraudAnalysis.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mobile Money Fraud Detection API
 * OpenAPI spec version: 0.1.0
 */
import type { FeatureImportance } from "./featureImportance";
import type { FraudAnalysisFraudLabel } from "./fraudAnalysisFraudLabel";

export interface FraudAnalysis {
  /** Risk score 0-100 */
  riskScore: number;
  isFraud: boolean;
  fraudLabel: FraudAnalysisFraudLabel;
  triggeredRules: string[];
  featureImportance: FeatureImportance[];
  explanation: string;
  modelUsed: string;
  confidence: number;
}