# lib/api-zod/src/generated/types/modelMetrics.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mobile Money Fraud Detection API
 * OpenAPI spec version: 0.1.0
 */

export interface ModelMetrics {
  precision: number;
  recall: number;
  f1Score: number;
  accuracy: number;
  specificity: number;
  falsePositiveRate: number;
  falseNegativeRate: number;
  truePositives: number;
  trueNegatives: number;
  falsePositives: number;
  falseNegatives: number;
  auc: number;
}