# lib/api-zod/src/generated/types/volumeDataPoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mobile Money Fraud Detection API
 * OpenAPI spec version: 0.1.0
 */

export interface VolumeDataPoint {
  time: string;
  total: number;
  fraudulent: number;
  suspicious: number;
  clean: number;
}